Enable expect for automatic and manual installation
Includes an update of original autoinstall.expect commands for new gui
and updates to validate_release.py and helper scripts/config files to
pull the latest version of Clear Linux.

The new expect script runs through the manual installation steps in the
ister gui. The script is run via the same validate_release.py script as
the automatic installation script.

+++ update_gui_expect.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/bash
-# usage: update_gui_expect.sh <disk image>
+# usage: update_gui_expect.sh <disk image> <expect file> <expect service>
 
 function usage()
 {
-    echo "Usage: $0 <disk.img>"
+    echo "Usage: $0 <disk.img> <file.expect> <expect.service>"
     exit 1
 }
 
-if [ $# -ne 1 ]
+if [ $# -ne 3 ]
 then
         usage
 fi
@@ -23,13 +23,25 @@ then
         exit 1
 fi
 
+if [ ! -e $2 ]
+then
+        echo $2: Does not exist
+        exit 1
+fi
+
+if [ ! -e $3 ]
+then
+        echo $3: Does not exist
+        exit 1
+fi
+
 mnt=$(/usr/bin/mktemp -d)
 next_dev=$(sudo losetup -f --show -P $1)
 sudo /usr/bin/mount ${next_dev}p2 $mnt
 sudo cp ister_gui.py ${mnt}/usr/bin/ister_gui.py
 sudo cp ister.py ${mnt}/usr/bin/ister.py
-sudo cp autoinstall.expect ${mnt}/usr/bin
-sudo cp ister-expect.service ${mnt}/usr/lib/systemd/system/ister.service
+sudo cp $2 ${mnt}/usr/bin
+sudo cp $3 ${mnt}/usr/lib/systemd/system/ister.service
 sudo /usr/bin/umount $mnt
 sudo /usr/bin/losetup -D
 echo "$1 set to use expect to drive install"
